embedding-service/requirements.txt
torch==2.6.0
transformers==4.49.0
pillow==11.1.0
kafka-python==2.0.6
python-dotenv==1.0.1
flask==3.1.0
requests==2.32.3